move utils to dedicated folder
